backend/requirements.txt
fastapi
uvicorn
python-multipart
onnxruntime
numpy
scikit-image
requests
Pillow
